+++ docs/custom_datasources_index.rst
@@ -211,10 +211,10 @@ insights.specs.datasources.sys_fs_cgroup_memory_tasks_number
     :show-inheritance:
     :undoc-members:
 
-insights.specs.datasources.system_user_pkgs
+insights.specs.datasources.rpm_pkgs
 -------------------------------------------
 
-.. automodule:: insights.specs.datasources.system_user_pkgs
+.. automodule:: insights.specs.datasources.rpm_pkgs
     :members: pkgs_with_writable_dirs
     :show-inheritance:
     :undoc-members:
