+++ docs/shared_parsers_catalog/rpm_pkgs.rst
@@ -1,0 +1,3 @@
+.. automodule:: insights.parsers.rpm_pkgs
+   :members:
+   :show-inheritance:
