--- docs/shared_parsers_catalog/system_user_pkgs.rst
@@ -1,3 +1,0 @@
-.. automodule:: insights.parsers.system_user_pkgs
-   :members:
-   :show-inheritance:
