+++ insights/specs/__init__.py
@@ -699,7 +699,7 @@ class Specs(SpecSet):
     sysctl_conf_initramfs = RegistryPoint(multi_output=True)
     sysctl_d_conf_etc = RegistryPoint(multi_output=True)
     sysctl_d_conf_usr = RegistryPoint(multi_output=True)
-    system_user_pkgs = RegistryPoint()
+    rpm_pkgs = RegistryPoint()
     systemctl_cat_dnsmasq_service = RegistryPoint()
     systemctl_cat_rpcbind_socket = RegistryPoint()
     systemctl_cinder_volume = RegistryPoint()
